+++ CHANGELOG.md
@@ -56,6 +56,8 @@ IMPROVEMENTS:
 * Import functions can now use a custom separator instead of "." [#343]
 * `resource/vcd_catalog_media` Add computed properties (`is_iso`, `owner_name`, `is_published`, `creation_date`, `size`, `status`, `storage_profile_name`) and full read support [#340]
 * `resource/vcd_catalog_media` MediaItem state ID changed from colon separated list of catalog name and media name to vCD ID [#340]
+* Import functions can now use custom separators instead of "." [#343]
+* `resource/vcd_vapp_vm` has new field `computer_name` [#334] [#347]
 
 BUG FIXES:
 * Change default value for `vcd_org.deployed_vm_quota` and `vcd_org.stored_vm_quota`. It was incorrectly set at `-1` instead of `0`.